Make it work by creating Q&A and .html files
To make it work i created some data in .JS file which contains questions
and answers. This data is overwriting data on existing form basing on
name of current .html page.

I know that this kind of stuff should be made rather on backend side,
however i want to practise with JS.

Due to fact that all .html files are the same (differences are made by
.js) it probably should be updated to using querystring and redirecting
to the same page each time form is submited (instead of creating
multiply .html files)